Put Tree details in custom component
